WindowsFontInstaller: support .fon and .ttc, proper path handling, cmd-like prints, retry-logic, fixed registry bugs

+++ FontRegister/Core/WindowsFontInstaller.cs
@@ -9,6 +9,8 @@ namespace FontRegister;
 
 public class WindowsFontInstaller : IFontInstaller
 {
+    private static readonly string[] _supportedExtensions = { ".ttf", ".otf", ".fon", ".ttc" };
+
     [DllImport("gdi32.dll")]
     private static extern int AddFontResourceA(string lpszFilename);
 
@@ -26,29 +28,69 @@ public class WindowsFontInstaller : IFontInstaller
         _systemNotifier = systemNotifier;
     }
 
-    public void InstallFont(string fontPath)
+    public bool InstallFont(string fontPath)
     {
+        string? fileName = null;
         try
         {
-            string fileName = Path.GetFileName(fontPath);
-            string fileExtension = Path.GetExtension(fontPath).ToLower();
-            string fontName = Path.GetFileNameWithoutExtension(fontPath);
+            fileName = Path.GetFileName(fontPath);
+            fontPath = Path.GetFullPath(fontPath).Replace("/", "\\"); //normalize
+            fontPath = Path.ChangeExtension(fontPath, Path.GetExtension(fontPath)?.ToLower());
+
+            if (!File.Exists(fontPath))
+            {
+                Console.WriteLine($"{fontPath}: Font file path not found.");
+                return false;
+            }
+
+            var fileExtension = Path.GetExtension(fontPath);
+            var fontName = Path.GetFileNameWithoutExtension(fontPath);
+            fontName = char.ToUpper(fontName[0]) + fontName.Substring(1); //first letter capital
+
+            var localFontDir = GetLocalFontDirectory();
+
+            //check if font already installed, our normalized version vs given version
+            if (File.Exists(Path.Combine(localFontDir, fileName)) || File.Exists(Path.Combine(localFontDir, Path.GetFileName(fontPath)))) 
+            {
+                Console.WriteLine($"{fileName}: Font already installed.");
+                return false;
+            }
 
             // Adjust font name based on file type
-            string registryFontName = fileExtension == ".otf" //similar behavior in default windows font installer
-                ? $"{fontName} (OpenType)"
-                : fontName;
+            var registryFontName = fileExtension switch
+            {
+                //similar behavior in default windows font installer
+                ".otf" => $"{fontName} (OpenType)",
+                ".ttc" => $"{fontName} (TrueType)",
+                ".fon" => $"{fontName} (VGA res)",
+                _ => fontName
+            };
 
-            string localFontDir = GetLocalFontDirectory();
-            string destPath = Path.Combine(localFontDir, fileName);
+            var destPath = Path.Combine(localFontDir, fileName);
 
             // Copy the font file
-            File.Copy(fontPath, destPath, true);
+            int copyAttempts = 0;
+            while (true)
+            {
+                try
+                {
+                    copyAttempts++;
+                    if (!File.Exists(destPath))
+                        File.Copy(fontPath, destPath, overwrite: true);
+                    break;
+                }
+                catch (Exception e)
+                {
+                    Thread.Sleep(100);
+                    if (copyAttempts > 10)
+                        throw new InvalidOperationException($"{fileName}: Failed to copy font file: {e.Message}");
+                }
+            }
 
             // Add the font resource
             if (AddFontResourceA(destPath) == 0)
             {
-                throw new InvalidOperationException("Failed to add font resource.");
+                throw new InvalidOperationException($"{fileName}: Failed to add font resource.");
             }
 
             using (var currentVersion = Registry.CurrentUser.OpenSubKey(@"Software\Microsoft\Windows NT\CurrentVersion"))
@@ -60,48 +102,82 @@ public class WindowsFontInstaller : IFontInstaller
             }
 
             // Add to current user registry
-            using (RegistryKey fontsKey = Registry.CurrentUser.OpenSubKey(@"Software\Microsoft\Windows NT\CurrentVersion\Fonts", true)!)
+            using (var fontsKey = Registry.CurrentUser.OpenSubKey(@"Software\Microsoft\Windows NT\CurrentVersion\Fonts", true)!)
             {
                 if (fontsKey == null)
                 {
-                    throw new InvalidOperationException("Unable to open the fonts registry key.");
+                    throw new InvalidOperationException($"{fileName}: Unable to open the fonts registry key.");
                 }
 
-                fontsKey.SetValue(registryFontName, destPath);
+                fontsKey.SetValue(registryFontName, fontPath);
             }
 
-            Console.WriteLine($"Font {registryFontName} installed successfully.");
+            Console.WriteLine($"{fileName}: Font installed successfully.");
 
             _systemNotifier?.NotifyFontChange();
+            return true;
         }
         catch (Exception ex)
         {
-            Console.WriteLine($"Error installing font {Path.GetFileName(fontPath)}: {ex}");
+            Console.WriteLine($"{fileName ?? fontPath}: Error installing font: {ex}");
+            return false;
         }
     }
 
-    public void UninstallFont(string fontName)
+    public bool UninstallFont(string fontNameOrPath)
     {
+        string? fileName = null; //the file name with extension
+        string? fontPath = null; //the full path of the font file in local font directory
         try
         {
-            string localFontDir = GetLocalFontDirectory();
-            string[] possibleExtensions = { ".ttf", ".otf", ".fon", ".ttc" };
-            string fontPath = null;
+            if (fontNameOrPath.Contains("\\") || fontNameOrPath.Contains("/") || Path.IsPathRooted(fontNameOrPath) || fontNameOrPath.Contains(".."))
+                fontNameOrPath = Path.GetFullPath(fontNameOrPath).Replace("/", "\\"); //normalize
 
-            foreach (var ext in possibleExtensions)
+            var localFontDir = GetLocalFontDirectory();
+            //handle full path inside local font directory passed
+            if (Path.IsPathRooted(fontNameOrPath))
             {
-                string possiblePath = Path.Combine(localFontDir, fontName + ext);
-                if (File.Exists(possiblePath))
+                fileName = Path.GetFileName(fontNameOrPath);
+                fileName = Path.ChangeExtension(fileName, Path.GetExtension(fileName)?.ToLower());
+                fontNameOrPath = Path.GetFullPath(fontNameOrPath).Replace("/", "\\"); //normalize
+                if (!fontNameOrPath.StartsWith(localFontDir, StringComparison.OrdinalIgnoreCase))
+                    throw new InvalidOperationException($"{fileName}: Cannot uninstall fonts outside the local font directory, path: {fontNameOrPath}, expected: {Path.Combine(localFontDir, fontNameOrPath)}");
+            }
+            else
+            {
+                //handle just font name passed, search in local font directory for all supported extensions
+                if (!Path.HasExtension(fontNameOrPath))
                 {
-                    fontPath = possiblePath;
-                    break;
+                    fileName = fontNameOrPath; //for error message, expected to be replaced
+                    foreach (var ext in _supportedExtensions)
+                    {
+                        var potentialFileName = fontNameOrPath + ext;
+                        var potentialPath = Path.Combine(localFontDir, potentialFileName);
+                        if (File.Exists(potentialPath))
+                        {
+                            if (fontPath != null)
+                                throw new InvalidOperationException($"{fileName}: Multiple font files found with the same name but different extensions ({Path.GetExtension(fontPath)}, {ext}). Please specify extension.");
+
+                            fontPath = potentialPath;
+                            fileName = potentialFileName;
+                        }
+                    }
+                }
+                else
+                {
+                    //handle font name with extension passed
+                    fileName = fontNameOrPath;
+                    fileName = Path.ChangeExtension(fileName, Path.GetExtension(fileName)?.ToLower());
+                    fontPath = Path.Combine(localFontDir, fontNameOrPath);
+                    if (!_supportedExtensions.Contains(Path.GetExtension(fontNameOrPath), StringComparer.OrdinalIgnoreCase))
+                        throw new InvalidOperationException($"{fileName}: Unsupported font extension: {Path.GetExtension(fontNameOrPath)}");
                 }
             }
 
-            if (fontPath == null)
+            if (fontPath == null || !File.Exists(fontPath))
             {
-                Console.WriteLine($"Font {fontName} not found.");
-                return;
+                Console.WriteLine($"{fileName}: Font not found.");
+                return false;
             }
 
             // Remove the font resource
@@ -109,13 +185,33 @@ public class WindowsFontInstaller : IFontInstaller
             {
                 //read error
                 var error = Marshal.GetLastWin32Error();
-                throw new InvalidOperationException("Failed to remove font resource. (" + error + ")");
+                throw new InvalidOperationException($"{fileName}: Failed to remove font resource. (errorcode: {error})");
             }
 
             // Delete the font file
-            File.Delete(fontPath);
+            int deletionAttempts = 0;
+            while (true)
+            {
+                try
+                {
+                    deletionAttempts++;
+                    if (File.Exists(fontPath))
+                        File.Delete(fontPath);
+                    break;
+                }
+                catch (DirectoryNotFoundException)
+                {
+                    break;
+                }
+                catch (Exception e)
+                {
+                    Thread.Sleep(100);
+                    if (deletionAttempts > 10)
+                        throw new InvalidOperationException($"{fileName}: Failed to delete font file: {e.Message}");
+                }
+            }
 
-            // Remove from current user registry
+            // Ensure fonts registry key exists
             using (var currentVersion = Registry.CurrentUser.OpenSubKey(@"Software\Microsoft\Windows NT\CurrentVersion"))
             {
                 if (!currentVersion!.GetSubKeyNames().Contains("Fonts"))
@@ -124,15 +220,26 @@ public class WindowsFontInstaller : IFontInstaller
                 }
             }
 
-            using (RegistryKey fontsKey = Registry.CurrentUser.OpenSubKey(@"Software\Microsoft\Windows NT\CurrentVersion\Fonts", true)!)
+            // Remove from current user registry
+            using (var fontsKey = Registry.CurrentUser.OpenSubKey(@"Software\Microsoft\Windows NT\CurrentVersion\Fonts", true)!)
             {
                 if (fontsKey == null)
-                {
-                    throw new InvalidOperationException("Unable to open the fonts registry key.");
-                }
+                    throw new InvalidOperationException($"{fileName}: Unable to open the fonts registry key.");
 
-                string registryValueName = fontsKey.GetValueNames()
-                    .FirstOrDefault(name => name.StartsWith(fontName, StringComparison.OrdinalIgnoreCase));
+                var installedFontsValueNames = fontsKey.GetValueNames();
+
+                // Attempt deleting when registry value is just the file name or full path
+                var registryValueName = installedFontsValueNames
+                    .FirstOrDefault(regName =>
+                    {
+                        if (Path.IsPathRooted(regName))
+                            regName = Path.GetFullPath(regName);
+                        if (regName.Contains("\\"))
+                            regName = regName.Replace("/", "\\");
+
+                        return regName.Equals(fontPath, StringComparison.OrdinalIgnoreCase) ||
+                               regName.ToLower().Contains(fileName.ToLower());
+                    });
 
                 if (registryValueName != null)
                 {
@@ -140,29 +247,31 @@ public class WindowsFontInstaller : IFontInstaller
                 }
                 else
                 {
-                    Console.WriteLine($"Font {fontName} not found in registry.");
+                    Console.WriteLine($"{fileName}: Font not found in registry.");
                 }
             }
 
-            Console.WriteLine($"Font {fontName} uninstalled successfully.");
+            Console.WriteLine($"{fileName}: Font uninstalled successfully.");
 
             _systemNotifier?.NotifyFontChange();
+            return true;
         }
         catch (Exception ex)
         {
-            Console.WriteLine($"Error uninstalling font {fontName}: {ex.Message}");
+            Console.WriteLine($"{fileName}: Error uninstalling font: {ex.Message}");
+            return false;
         }
     }
 
     private string GetLocalFontDirectory()
     {
-        string localFontDir = Path.Combine(
+        var localFontDir = Path.Combine(
             Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData),
             "Microsoft", "Windows", "Fonts"
         );
 
         //normalize path
-        localFontDir = Path.GetFullPath(localFontDir);
+        localFontDir = Path.GetFullPath(localFontDir).Replace("/", "\\");
 
         Directory.CreateDirectory(localFontDir);
         return localFontDir;
